+++ Common/include/Image.hpp
@@ -5,7 +5,7 @@
 
 #include <memory>
 
-class Pixel;
+#include "Pixel.hpp"
 
 class Image
 {
@@ -13,7 +13,7 @@ public:
 	enum class StartPosition : uint8_t { TopLeft, BottomLeft };
 
 	Image(uint32_t width, uint32_t height, StartPosition startPosition = StartPosition::TopLeft);
-	~Image();
+	virtual ~Image() = default;
 
 	uint32_t GetWidth() const;
 	uint32_t GetHeight() const;
@@ -24,14 +24,14 @@ public:
 	inline StartPosition GetStartPosition() const noexcept { return this->startPosition; };
 
 protected:
-	Image();
+	Image() = default;
 
 	StartPosition startPosition = StartPosition::TopLeft;
 
-	uint32_t width;
-	uint32_t height;
-
 	std::unique_ptr<std::unique_ptr<Pixel>[]> image;
+
+	uint32_t width  = 0;
+	uint32_t height = 0;
 
 };
