+++ Common/include/png/Adler32.hpp
@@ -9,6 +9,9 @@
 class Adler32
 {
 public:
+	Adler32() = delete;
+	~Adler32() = delete;
+
 	static uint32_t Decode(const std::vector<uint8_t> &data);
 
 };
